skitter_api.v0_crawl: ensure we have a new response to return
This removes the v0_cache usage from v0_crawl, and ensures that the
response we're attempting to return was actually crawled _after_ we got
the request to do so. This fixes the rare case where someone is munging
directly with the db and makes RemoteWebScraper.scrape return early
without an exception.